# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SIMA API",
        "description": "Sistem Informasi Manajemen Amanah — REST API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API base path"
        }
    ],
    "paths": {
        "/accounts": {
            "get": {
                "tags": [
                    "Account"
                ],
                "summary": "Daftar akun kas/bank",
                "operationId": "96daed062b209799ebd8c7e84dc97aed",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiEnvelope"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Account"
                ],
                "summary": "Buat akun kas/bank",
                "operationId": "b3234b57f04a5c23b5d70ac83678df8c",
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiEnvelope"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/accounts/{account}": {
            "get": {
                "tags": [
                    "Account"
                ],
                "summary": "Detail akun kas/bank",
                "operationId": "fa80b13ea2ab503138c7faf25fcf984f",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiEnvelope"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Account"
                ],
                "summary": "Ubah akun kas/bank",
                "operationId": "e5a4f19af009092c65acd1e4232c8f5b",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiEnvelope"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Account"
                ],
                "summary": "Nonaktifkan akun kas/bank",
                "operationId": "0fcbfcc738e1e05609659d9a2bdde35d",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiEnvelope"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/attachments": {
            "get": {
                "tags": [
                    "Attachment"
                ],
                "summary": "Daftar lampiran",
                "operationId": "b5a27a67ec05dc2fb0f76c40e348e529",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiEnvelope"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Attachment"
                ],
                "summary": "Unggah lampiran",
                "operationId": "5685c13f391a59a7e8ac0114b0c0130c",
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiEnvelope"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/attachments/{attachment}/download": {
            "get": {
                "tags": [
                    "Attachment"
                ],
                "summary": "Unduh lampiran",
                "operationId": "d4c1160abd27e2b8ee8c6140f4cbbc42",
                "responses": {
                    "200": {
                        "description": "File stream"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/attachments/{attachment}": {
            "delete": {
                "tags": [
                    "Attachment"
                ],
                "summary": "Hapus lampiran",
                "operationId": "dd1cf9e7079b3f23f356fa1ece5852dd",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiEnvelope"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/audits": {
            "get": {
                "tags": [
                    "Audit"
                ],
                "summary": "Daftar audit trail",
                "operationId": "194a055698c38ffe50c8013ead50d35d",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiEnvelope"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/audits/{audit}": {
            "get": {
                "tags": [
                    "Audit"
                ],
                "summary": "Detail audit trail",
                "operationId": "7b1d3da95a8f7c91e1d3a1d4f8bf9d82",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiEnvelope"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "ApiEnvelope": {
                "properties": {
                    "success": {
                        "type": "boolean"
                    },
                    "message": {
                        "type": "string",
                        "nullable": true
                    },
                    "data": {
                        "nullable": true
                    },
                    "meta": {
                        "nullable": true
                    },
                    "errors": {
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Token dari POST /api/login",
                "bearerFormat": "Sanctum Token",
                "scheme": "bearer"
            }
        }
    },
...
}